# public/openapi.json
{"openapi":"3.0.0","info":{"title":"@tscircuit/layout-tester","version":"1.0.0"},"servers":[{"url":"https://example.com"}],"tags":[],"paths":{"layout/list":{"post":{"summary":"layout/list","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"layouts":{"type":"array","items":{"type":"object","properties":{"layout_name":{"type":"string"},"layout_group_name":{"type":"string"},"layout":{"type":"object","properties":{},"required":[]}},"required":["layout_name","layout_group_name","layout"]}},"ok":{"type":"boolean"}},"required":["layouts","ok"]}}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"layout_group_name":{"type":"string"}},"required":[]}}}},"tags":[]},"get":{"summary":"layout/list","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"layouts":{"type":"array","items":{"type":"object","properties":{"layout_name":{"type":"string"},"layout_group_name":{"type":"string"},"layout":{"type":"object","properties":{},"required":[]}},"required":["layout_name","layout_group_name","layout"]}},"ok":{"type":"boolean"}},"required":["layouts","ok"]}}}},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"layout_group_name":{"type":"string"}},"required":[]}}}},"tags":[]}},"layout/update":{"patch":{"summary":"layout/update","responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"layout_group_name":{"type":"string"},"layout_name":{"type":"string"},"layout":{"type":"object","properties":{},"required":[]}},"required":["layout_group_name","layout_name","layout"]}}}},"tags":[]},"post":{"summary":"layout/update","responses":{"200":{"description":"OK"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"layout_group_name":{"type":"string"},"layout_name":{"type":"string"},"layout":{"type":"object","properties":{},"required":[]}},"required":["layout_group_name","layout_name","layout"]}}}},"tags":[]}}},"components":{"securitySchemes":{}}}